refactor(Balance): AND-1216 Use BalanceReporter adapters and more use of CryptoValue over Long/BigInteger/BigDecimal (#1011)

+++ app/src/main/java/piuk/blockchain/android/data/balance/adapters/BitcoinBalanceReportAdapter.kt
@@ -21,5 +21,5 @@ private class BitcoinBalanceReportAdapter(
         payloadManager.getAddressBalance(address).toBalance()
 
     private fun BigInteger.toBalance() =
-        CryptoValue(CryptoCurrency.BTC, this.toLong())
+        CryptoValue(CryptoCurrency.BTC, this)
 }
